+++ nanomesh/__init__.py
@@ -1,8 +1,10 @@
-# -*- coding: utf-8 -*-
-"""Documentation about nanomesh."""
 import logging
 import sys
 
+from .mesh2d import Mesher2D
+from .mesh3d import Mesher3D
+from .mesh_containers import TetraMesh, TriangleMesh
+from .mesh_utils import simple_triangulate
 from .plane import Plane
 from .volume import Volume
 
@@ -18,6 +20,11 @@ __all__ = [
     '__author__',
     '__email__',
     '__version__',
+    'Mesher2D',
+    'Mesher3D',
     'Plane',
+    'simple_triangulate',
+    'TetraMesh',
+    'TriangleMesh',
     'Volume',
 ]
